Forgot to rename asset paths.
Go figure that these aren't included by default.
