+++ src/renderer/src/components/moves/MoveTable.tsx
@@ -3,61 +3,61 @@ import DoubleIcon from './DoubleIcon'
 import SimpleIcon from '../settings/SimpleIcon'
 import SlamSelector from './SlamSelector'
 
-import chunkyIcon from '../assets/images/chunky.png'
-import chunkyGunIcon from '../assets/images/chunky_gun.png'
-import chunkyInstIcon from '../assets/images/chunky_inst.png'
-import chunkyMoveIcon from '../assets/images/chunky_move.png'
-import chunkyBarrelIcon from '../assets/images/chunkybarrel.png'
-import chunkyPadIcon from '../assets/images/chunkypad.png'
-import diddyIcon from '../assets/images/diddy.png'
-import diddyGunIcon from '../assets/images/diddy_gun.png'
-import diddyInstIcon from '../assets/images/diddy_inst.png'
-import diddyMoveIcon from '../assets/images/diddy_move.png'
-import diddyBarrelIcon from '../assets/images/diddybarrel.png'
-import diddyPadIcon from '../assets/images/diddypad.png'
-import dkIcon from '../assets/images/dk.png'
-import dkGunIcon from '../assets/images/dk_gun.png'
-import dkInstIcon from '../assets/images/dk_inst.png'
-import dkMoveIcon from '../assets/images/dk_move.png'
-import dkBarrelIcon from '../assets/images/dkbarrel.png'
-import dkPadIcon from '../assets/images/dkpad.png'
-import lankyIcon from '../assets/images/lanky.png'
-import lankyGunIcon from '../assets/images/lanky_gun.png'
-import lankyInstIcon from '../assets/images/lanky_inst.png'
-import lankyMoveIcon from '../assets/images/lanky_move.png'
-import lankyBarrelIcon from '../assets/images/lankybarrel.png'
-import lankyPadIcon from '../assets/images/lankypad.png'
-import tinyIcon from '../assets/images/tiny.png'
-import tinyGunIcon from '../assets/images/tiny_gun.png'
-import tinyInstIcon from '../assets/images/tiny_inst.png'
-import tinyMoveIcon from '../assets/images/tiny_move.png'
-import tinyBarrelIcon from '../assets/images/tinybarrel.png'
-import tinyPadIcon from '../assets/images/tinypad.png'
+import chunkyIcon from '../../assets/images/chunky.png'
+import chunkyGunIcon from '../../assets/images/chunky_gun.png'
+import chunkyInstIcon from '../../assets/images/chunky_inst.png'
+import chunkyMoveIcon from '../../assets/images/chunky_move.png'
+import chunkyBarrelIcon from '../../assets/images/chunkybarrel.png'
+import chunkyPadIcon from '../../assets/images/chunkypad.png'
+import diddyIcon from '../../assets/images/diddy.png'
+import diddyGunIcon from '../../assets/images/diddy_gun.png'
+import diddyInstIcon from '../../assets/images/diddy_inst.png'
+import diddyMoveIcon from '../../assets/images/diddy_move.png'
+import diddyBarrelIcon from '../../assets/images/diddybarrel.png'
+import diddyPadIcon from '../../assets/images/diddypad.png'
+import dkIcon from '../../assets/images/dk.png'
+import dkGunIcon from '../../assets/images/dk_gun.png'
+import dkInstIcon from '../../assets/images/dk_inst.png'
+import dkMoveIcon from '../../assets/images/dk_move.png'
+import dkBarrelIcon from '../../assets/images/dkbarrel.png'
+import dkPadIcon from '../../assets/images/dkpad.png'
+import lankyIcon from '../../assets/images/lanky.png'
+import lankyGunIcon from '../../assets/images/lanky_gun.png'
+import lankyInstIcon from '../../assets/images/lanky_inst.png'
+import lankyMoveIcon from '../../assets/images/lanky_move.png'
+import lankyBarrelIcon from '../../assets/images/lankybarrel.png'
+import lankyPadIcon from '../../assets/images/lankypad.png'
+import tinyIcon from '../../assets/images/tiny.png'
+import tinyGunIcon from '../../assets/images/tiny_gun.png'
+import tinyInstIcon from '../../assets/images/tiny_inst.png'
+import tinyMoveIcon from '../../assets/images/tiny_move.png'
+import tinyBarrelIcon from '../../assets/images/tinybarrel.png'
+import tinyPadIcon from '../../assets/images/tinypad.png'
 
-import barrelIcon from '../assets/images/barrel2.png'
-import diveIcon from '../assets/images/diving.png'
-import orangeIcon from '../assets/images/orange.png'
-import vineIcon from '../assets/images/vine.png'
+import barrelIcon from '../../assets/images/barrel2.png'
+import diveIcon from '../../assets/images/diving.png'
+import orangeIcon from '../../assets/images/orange.png'
+import vineIcon from '../../assets/images/vine.png'
 
-import filmWaveBoth from '../assets/images/filmwave.png'
-import filmWaveLeft from '../assets/images/filmwave1.png'
-import filmWaveRight from '../assets/images/filmwave2.png'
+import filmWaveBoth from '../../assets/images/filmwave.png'
+import filmWaveLeft from '../../assets/images/filmwave1.png'
+import filmWaveRight from '../../assets/images/filmwave2.png'
 
-import homeScopeBoth from '../assets/images/homingscope.png'
-import homeScopeLeft from '../assets/images/homingscope1.png'
-import homeScopeRight from '../assets/images/homingscope2.png'
+import homeScopeBoth from '../../assets/images/homingscope.png'
+import homeScopeLeft from '../../assets/images/homingscope1.png'
+import homeScopeRight from '../../assets/images/homingscope2.png'
 
-import companyCoinBoth from '../assets/images/n64rw_coin.png'
-import companyCoinRight from '../assets/images/n64rw_coin_n64bw.png'
-import companyCoinLeft from '../assets/images/n64rw_coin_rwbw.png'
+import companyCoinBoth from '../../assets/images/n64rw_coin.png'
+import companyCoinRight from '../../assets/images/n64rw_coin_n64bw.png'
+import companyCoinLeft from '../../assets/images/n64rw_coin_rwbw.png'
 
-import goldBananaIcon from '../assets/images/GB.png'
-import bananaMedalIcon from '../assets/images/bananamedal.png'
-import beanIcon from '../assets/images/bean.png'
-import crownIcon from '../assets/images/crown.png'
-import fairyIcon from '../assets/images/fairy.png'
-import pearlIcon from '../assets/images/pearl.png'
-import rainbowCoinICon from '../assets/images/rainbowcoin.png'
+import goldBananaIcon from '../../assets/images/GB.png'
+import bananaMedalIcon from '../../assets/images/bananamedal.png'
+import beanIcon from '../../assets/images/bean.png'
+import crownIcon from '../../assets/images/crown.png'
+import fairyIcon from '../../assets/images/fairy.png'
+import pearlIcon from '../../assets/images/pearl.png'
+import rainbowCoinICon from '../../assets/images/rainbowcoin.png'
 import useDonkStore from '@renderer/store'
 import { useShallow } from 'zustand/react/shallow'
