+++ src/renderer/src/components/settings/BananaportSelector.tsx
@@ -4,8 +4,8 @@ import { useShallow } from 'zustand/react/shallow'
 import useDonkStore from '@renderer/store'
 import { BananaportRange } from '@renderer/store/common'
 
-import portFullIcon from '../assets/images/settings/bananaport.png'
-import portHalfIcon from '../assets/images/settings/bananaport_isles.png'
+import portFullIcon from '../../assets/images/settings/bananaport.png'
+import portHalfIcon from '../../assets/images/settings/bananaport_isles.png'
 
 const portToIcon = (num: BananaportRange): string => {
   return num == 1 ? portHalfIcon : portFullIcon
