+++ src/renderer/src/components/settings/GeneratorSettings.tsx
@@ -8,33 +8,33 @@ import CountSelector from './CountSelector'
 import HelmSelector from './HelmSelector'
 import SimpleIcon from './SimpleIcon'
 
-import beanIcon from '../assets/images/bean.png'
-import crownIcon from '../assets/images/crown.png'
-import peanutIcon from '../assets/images/diddy_gun.png'
-import guitarIcon from '../assets/images/diddy_inst.png'
-import rocketIcon from '../assets/images/diddybarrel.png'
-import coconutIcon from '../assets/images/dk_gun.png'
-import grabIcon from '../assets/images/dk_move.png'
-import fairyIcon from '../assets/images/fairy.png'
-import keyIcon from '../assets/images/key.png'
-import blueprintIcon from '../assets/images/lanky_bp.png'
-import grapeIcon from '../assets/images/lanky_gun.png'
-import companyCoinIcon from '../assets/images/n64rw_coin.png'
-import rainbowCoinIcon from '../assets/images/rainbowcoin.png'
-import arenaIcon from '../assets/images/settings/arena.png'
-import bananaMedalIcon from '../assets/images/settings/bananamedal.gif'
-import bonusIcon from '../assets/images/settings/bonus.png'
-import bananaIcon from '../assets/images/settings/bunch.png'
-import dirtIcon from '../assets/images/settings/dirt_patch.png'
-import shopIcon from '../assets/images/settings/funky.png'
-import hardShootIcon from '../assets/images/settings/hard_shoot.png'
-import kasplatIcon from '../assets/images/settings/kasplat.png'
-import melonIcon from '../assets/images/settings/melon.png'
-import toughBananaIcon from '../assets/images/settings/rabbit.png'
-import slamSwitchIcon from '../assets/images/settings/slam_switch.png'
-import switchsanityIcon from '../assets/images/settings/switch.png'
-import slamIcon from '../assets/images/slam1.png'
-import featherIcon from '../assets/images/tiny_gun.png'
+import beanIcon from '../../assets/images/bean.png'
+import crownIcon from '../../assets/images/crown.png'
+import peanutIcon from '../../assets/images/diddy_gun.png'
+import guitarIcon from '../../assets/images/diddy_inst.png'
+import rocketIcon from '../../assets/images/diddybarrel.png'
+import coconutIcon from '../../assets/images/dk_gun.png'
+import grabIcon from '../../assets/images/dk_move.png'
+import fairyIcon from '../../assets/images/fairy.png'
+import keyIcon from '../../assets/images/key.png'
+import blueprintIcon from '../../assets/images/lanky_bp.png'
+import grapeIcon from '../../assets/images/lanky_gun.png'
+import companyCoinIcon from '../../assets/images/n64rw_coin.png'
+import rainbowCoinIcon from '../../assets/images/rainbowcoin.png'
+import arenaIcon from '../../assets/images/settings/arena.png'
+import bananaMedalIcon from '../../assets/images/settings/bananamedal.gif'
+import bonusIcon from '../../assets/images/settings/bonus.png'
+import bananaIcon from '../../assets/images/settings/bunch.png'
+import dirtIcon from '../../assets/images/settings/dirt_patch.png'
+import shopIcon from '../../assets/images/settings/funky.png'
+import hardShootIcon from '../../assets/images/settings/hard_shoot.png'
+import kasplatIcon from '../../assets/images/settings/kasplat.png'
+import melonIcon from '../../assets/images/settings/melon.png'
+import toughBananaIcon from '../../assets/images/settings/rabbit.png'
+import slamSwitchIcon from '../../assets/images/settings/slam_switch.png'
+import switchsanityIcon from '../../assets/images/settings/switch.png'
+import slamIcon from '../../assets/images/slam1.png'
+import featherIcon from '../../assets/images/tiny_gun.png'
 
 const customStyles: Modal.Styles = {
   content: {
