+++ src/renderer/src/components/settings/LevelSelector.tsx
@@ -4,15 +4,15 @@ import { useShallow } from 'zustand/react/shallow'
 import useDonkStore from '@renderer/store'
 import { Level } from '@renderer/store/common'
 
-import aztecIcon from '../assets/images/aztec.png'
-import castleIcon from '../assets/images/castle.png'
-import cavesIcon from '../assets/images/caves.png'
-import factoryIcon from '../assets/images/factory.png'
-import forestIcon from '../assets/images/forest.png'
-import galleonIcon from '../assets/images/galleon.png'
-import helmIcon from '../assets/images/helm.png'
-import japesIcon from '../assets/images/japes.png'
-import unknownIcon from '../assets/images/unknown.png'
+import aztecIcon from '../../assets/images/aztec.png'
+import castleIcon from '../../assets/images/castle.png'
+import cavesIcon from '../../assets/images/caves.png'
+import factoryIcon from '../../assets/images/factory.png'
+import forestIcon from '../../assets/images/forest.png'
+import galleonIcon from '../../assets/images/galleon.png'
+import helmIcon from '../../assets/images/helm.png'
+import japesIcon from '../../assets/images/japes.png'
+import unknownIcon from '../../assets/images/unknown.png'
 
 type LevelSelectorProps = {
   storeKey: string
