+++ src/renderer/src/components/settings/Switchsanity.tsx
@@ -3,27 +3,27 @@ import { useShallow } from 'zustand/react/shallow'
 
 import SwitchsanitySwitch from './SwitchsanitySwitch'
 
-import chunkyKongIcon from '../assets/images/chunky.png'
-import chunkyGunIcon from '../assets/images/chunky_gun.png'
-import chunkyMusicIcon from '../assets/images/chunky_inst.png'
-import chunkyPadIcon from '../assets/images/chunkypad.png'
-import diddyKongIcon from '../assets/images/diddy.png'
-import diddyGunIcon from '../assets/images/diddy_gun.png'
-import diddyMusicIcon from '../assets/images/diddy_inst.png'
-import diddyMoveIcon from '../assets/images/diddy_move.png'
-import dkKongIcon from '../assets/images/dk.png'
-import dkGunIcon from '../assets/images/dk_gun.png'
-import dkMusicIcon from '../assets/images/dk_inst.png'
-import dkMoveIcon from '../assets/images/dk_move.png'
-import dkPadIcon from '../assets/images/dkpad.png'
-import lankyKongIcon from '../assets/images/lanky.png'
-import lankyGunIcon from '../assets/images/lanky_gun.png'
-import lankyMusicIcon from '../assets/images/lanky_inst.png'
-import lankyPadIcon from '../assets/images/lankypad.png'
-import tinyKongIcon from '../assets/images/tiny.png'
-import tinyGunIcon from '../assets/images/tiny_gun.png'
-import tinyMusicIcon from '../assets/images/tiny_inst.png'
-import tinyPadIcon from '../assets/images/tinypad.png'
+import chunkyKongIcon from '../../assets/images/chunky.png'
+import chunkyGunIcon from '../../assets/images/chunky_gun.png'
+import chunkyMusicIcon from '../../assets/images/chunky_inst.png'
+import chunkyPadIcon from '../../assets/images/chunkypad.png'
+import diddyKongIcon from '../../assets/images/diddy.png'
+import diddyGunIcon from '../../assets/images/diddy_gun.png'
+import diddyMusicIcon from '../../assets/images/diddy_inst.png'
+import diddyMoveIcon from '../../assets/images/diddy_move.png'
+import dkKongIcon from '../../assets/images/dk.png'
+import dkGunIcon from '../../assets/images/dk_gun.png'
+import dkMusicIcon from '../../assets/images/dk_inst.png'
+import dkMoveIcon from '../../assets/images/dk_move.png'
+import dkPadIcon from '../../assets/images/dkpad.png'
+import lankyKongIcon from '../../assets/images/lanky.png'
+import lankyGunIcon from '../../assets/images/lanky_gun.png'
+import lankyMusicIcon from '../../assets/images/lanky_inst.png'
+import lankyPadIcon from '../../assets/images/lankypad.png'
+import tinyKongIcon from '../../assets/images/tiny.png'
+import tinyGunIcon from '../../assets/images/tiny_gun.png'
+import tinyMusicIcon from '../../assets/images/tiny_inst.png'
+import tinyPadIcon from '../../assets/images/tinypad.png'
 
 const allKong = [dkKongIcon, diddyKongIcon, lankyKongIcon, tinyKongIcon, chunkyKongIcon]
 const allGun = [dkGunIcon, diddyGunIcon, lankyGunIcon, tinyGunIcon, chunkyGunIcon]
